Fixed bug with losing focus on player selection.
For some reason (still unsure), when the player selection within the home
level was entered, the first action would leave it in an unrecoverable state
(if only using buttons). The PauseMenu lost its ability to grab focus, which
didn't help but didn't hurt. To fix, the controlling input was explicitly
detached from UI actions when the PlayerSelection panel is entered.